+++ app/src/main/java/com/dejssa/radioserver/provider/Providers.java
@@ -1,0 +1,18 @@
+package com.dejssa.radioserver.provider;
+
+import android.content.Context;
+
+public class Providers {
+    private Context context;
+
+    public StationProvider station;
+    public RadioProvider radio;
+
+    public Providers(Context context) {
+        this.context = context;
+    }
+
+    public Context getContext() {
+        return context;
+    }
+}